Fix a bug in the safe and unsafe `testEq` implementations.
I ran into this bug in Reopt and @robdockins fixed it. The problem is
that `TestEquality.testEquality`, which `testEq` is used to implement,
does not constrain the indices to be equal on the two functors being
compared. Probably the person who wrote the code was thinking of
`EqF`, which does constrain the indices to be equal, which would have
really made the missing cases impossible.