Add countNoncesGenerated for profiling
Number of nonces generated is a good proxy for expressions created in
Crucible, so this is a handy way to measure that.

This required a fair amount of surgery in how nonce generators are
defined, but it should make for improved performance in most cases since
'freshNonce' can now be inlined.