for functions for TraversableF[C]
It's customary to have for as a flipped traverse. This adds the
functions to TraversableF and TraversableFC. It also exposes
forF_ which was previously named forMF_ but wasn't actually
exported. I've updated the name to match current practices
while exporting it.

I've deprecated the older forMFC_ name as the forM name
typically indicates a legacy Monad constraint.