Use base-orphans to provide TestEquality instance for Compose
GHC 8.10 will introduce a `TestEquality` instance for `Compose`,
which clashes with the one that `parameterized-utils` currently
defines in `Data.Parameterized.Compose` as an orphan instance.
To achieve forward compatibility with 8.10, this patch removes
`parameterized-utils`' orphan instance and replaces it with the one
in `base-orphans`, the canonical home for orphan instances for types
in the `base` library. (Note that on 8.10 or later, this will simply
use the instance found in `base`—the orphan instance is only used on
older GHCs for backwards compatibility.)