add CtxFlatten type family and flattenAssignment
For some purposes, it is useful to forget the nested structure of a
context of contexts, flattening it into one large context.  This
commit introduces a type family CtxFlatten to capture this at the
type level.

At the value level, the relevant operation consists in flattening
an assignment of assignments over this context of contexts, into
a flat assignment.