Re-organize with internal libraries to hide safe/unsafe internals.
This hides the Safe/Unsafe implementations of Context which prevents a
client from improperly using the wrong one.

This also allows client code and tests to be written in a fully
generic manner.

The internal movement/splitting of files is somewhat unfortunate, but
Cabal will look for a local file to supply an import statement before
checking libraries, so the files had to be segregated physically as
well as by internal library build.