TraversableFCWithIndex: Indexed traversals, folds, and maps
Add indexed functors, folds, and traversals inspired by the indexed-traversable
(which is where the naming scheme comes from) and lens packages.

The current contents of the List and Assignment modules provide additional
motivation: List already implements and exports a minimal set of methods for
each of the added classes, but Assignment only implemented
'traverseWithIndex' (AKA 'itraverseFC'); this indicates that identifying this
abstraction enables code sharing and increased functionality between these
similar modules/datatypes.

Additionally, Assignment now gets 'ifoldMapFC', 'imapFC' and friends "for free",
because they have default implementations based on 'itraverseFC'.

At some point, it may make sense to implement similar classes for the *F family:
'FunctorF', 'FoldableF', and 'TraversableF'. However, 'MapF' is the only
'TraversableF' that has an instance of 'IndexF', so this isn't as clearly a
code-sharing win, though such classes may uncover useful, as of yet
unimplemented functions for that API as the *FC ones did for 'Assignment'.