Raise lower version bounds on th-abstraction to >=0.4.2
`Data.Parameterized.ClassesC` relies on all of its imports being `Safe` or
`Trustworthy`, but prior to `th-abstraction-0.4.2.0`, the modules in
`th-abstraction` were unsafe. This unsafety can leak up to
`parameterized-utils`, causing the build failure observed in #139.

The simplest fix is to require `th-abstraction-0.4.2.0` as the minimum, which
this patch accomplishes. That version has been out for nearly two years, so
this shouldn't cause any undue complications with downstream dependencies not
supporting it.

Fixes #139.