Make Some into a newtype around Some from package some
The `Some` type from package `some` is a newtype, and is correspondingly more
performant than our previous data-based encoding.

This change will require a major version bump and an easy but still mandatory
migration for downstream codebases. Downstream code will either need to switch
to using the eliminator, or using pattern synonyms. Example migrations to use
pattern synonyms can be seen in this commit.

This patch also uses GeneralizedNewtypeDeriving to derive the upstream
`Semigroup` and `Monoid` instances, which are a good demonstration of the
utility of sharing an implementation with a well-thought-out upstream package.