Only compare against matching types for structural TH.
When generating `structuralEquality` or `structuralTypeEquality`, the former is
used in an `Eq` scenario for `(==) :: a -> a -> Bool` and so if there is a GADT
definition like the following:]

    data T1 a where
       T1_Int :: Int -> T1 Int
       T1_Char :: Char -> T1 Char

then the `(==)` function will only admit `T1_Int` constructors or `T1_Char`
constructors but not both.  However, `TestEquality` will parameterize over the
GADT type parameter and does not require them to match.  As a result, when
generating the splice for comparison:

    \x y ->
      case x of
        T1_Int ->
          case y of
            T1_Int -> ...
            _ -> False    -- << HERE
        T1_Char ->
          case y of
            T1_Char -> ...
            _ -> False    -- << AND HERE

that splice is fine for `TestEquality`, but because the types must match for `Eq`
then the two wildcard clauses marked by HERE comments above are superfluous and
cause warnings/errors.  There may be other explicit matchers (e.g. if T1 also had
a `T1_Nat :: Nat -> T1 Int` constructor) which will be needed, but the wildcard
should only be present for testEquality.

This patch makes the adjustment to allow for the differences in the
structuralEquality and structuralTypeEquality definitions, along with some tests
to demonstrate/exercise those changes.