Avoid -Wunused-imports warning with base-4.20 (GHC 9.10)
`base-4.20` (bundled with GHC 9.10) re-exports `foldl'` from the `Prelude`.
This causes an `-Wunused-imports` warning when building
`Data.Parameterized.Map`, which separately imports `foldl'` from `Data.List`.

This patch makes the `Data.List` import qualified to avoid this issue and make
it possible to build the code warning-free on GHC 9.10 and older.