ci: Publish Haddocks to GitHub Pages
In between full Hackage releases, it would be nice to be able to read
haddocks for the latest commit on the default branch. This commit builds
the haddocks in CI, and publishes them to GitHub Pages.

In more detail, this commit breaks the `cabal haddock` step out of the
testing job into a separate job for the sake of controlling concurrency
and permissions, while reusing the cache from the build step to build
the haddocks.