ci: Move Haddocks to their own workflow
The previous setup was buggy. The `doc` job would fail on non-`master`
branches, because it didn't have enough permissions to access the GitHub
Pages "environment", even though if it were run it wouldn't attempt to
upload to it.

Also, use a separate cache for documentation builds. This is
appropriate, as we can and should build with `--disable-documentation`
to save space and time in the `test` job, but *not* build with that flag
for the docs. This discrepancy causes dependencies to be built
differently, hence separate caches.

Return to checking Haddock well-formedness in `test`.